Web crawler  stevens
Program runs for 60 minutes and crawls the stevens website's.

It created the directory folder and based on page hierarchy.

Also the email found on in entire session is captured in csv file

I was not able to implement the robotest parser for every request made  the parsers would return false.